Don't show browser log and other messages
We just want to show a summary table and don't concern ourselves with
the details of individual tests. We leave the job of reporting those
details to other reporters.
We assume those other reporters usually show the browser messages when
appropriate. If we also show them it leads to useless duplicated output.

Fixes #15