Add a `browserList` config option
It configures when the list of connected browsers is shown before the
summary table. The possible values are `'always'`, `'never'` or
`'ifneeded'`.
`'ifneeded'` shows the browser list only if the summary table isn't
empty and contains values for more than one browser.

The inspiration for this change came from @georgecrawford, thanks for
the idea!