Remove schedule for CodeQL Github Action
If there is no activity in the repository for 60 days, Github seems
to automatically disable scheduled actions. It than also doesn't
run on push/... anymore.
The idea behind the schedule was, that the CodeQL database/analysis
might change over time, leading to different results even if the code
in this repository itself didn't change. I'm not sure that would really
be the case. But anyway, if having the schedule leads to having the
action automatically disabled, it's better to not have it on schedule.